+++ image_generator/test_font/__init__.py
@@ -1,0 +1,1 @@
+from image_generator.test_font.core import createFontTest